Fix compile warning

+++ src/juci.cc
@@ -108,7 +108,7 @@ Application::Application() : Gtk::Application("no.sout.juci", Gio::APPLICATION_N
 }
 
 int Application::run(int argc, char **argv){
-  for(size_t i = 0; i < argc; i++){
+  for(int i = 0; i < argc; i++){
     args.emplace_back(argv[i]);
   }
   return Gtk::Application::run(argc, argv);
